+++ plugins/modules/zhmc_password_rule_list.py
@@ -105,8 +105,8 @@ options:
   _faked_session:
     description:
       - "An internal parameter used for testing the module."
-    required: false
     type: raw
+    required: false
     default: null
 """