+++ plugins/modules/zhmc_storage_group.py
@@ -193,8 +193,8 @@ options:
   _faked_session:
     description:
       - "An internal parameter used for testing the module."
-    required: false
     type: raw
+    required: false
     default: null
 """
 
@@ -711,7 +711,7 @@ def process_properties(cpc, storage_group, params):
                 format(sg_name, cpc.name, sg_cpc.name))
 
     # handle the other properties
-    input_props = params.get('properties', None)
+    input_props = params['properties']
     if input_props is None:
         input_props = {}
     for prop_name in input_props: