+++ plugins/modules/zhmc_user_list.py
@@ -104,8 +104,8 @@ options:
   _faked_session:
     description:
       - "An internal parameter used for testing the module."
-    required: false
     type: raw
+    required: false
     default: null
 """