+++ plugins/modules/zhmc_user_role.py
@@ -240,8 +240,8 @@ options:
   _faked_session:
     description:
       - "An internal parameter used for testing the module."
-    required: false
     type: raw
+    required: false
     default: null
 """
 
@@ -556,7 +556,7 @@ def process_properties(client, urole, params):
         pass
 
     # handle the other properties
-    input_props = params.get('properties', None)
+    input_props = params['properties']
     if input_props is None:
         input_props = {}