+++ plugins/modules/zhmc_virtual_function.py
@@ -166,8 +166,8 @@ options:
   _faked_session:
     description:
       - "An internal parameter used for testing the module."
-    required: false
     type: raw
+    required: false
     default: null
 """
 
@@ -351,7 +351,7 @@ def process_properties(partition, vfunction, params):
     adapter_name_art_name = 'adapter_name'
 
     # handle the other properties
-    input_props = params.get('properties', None)
+    input_props = params['properties']
     if input_props is None:
         input_props = {}
     for prop_name in input_props: